Namespace all ConfigLoaders (we're going to add more)
