+++ spec/resque_pool_spec.rb
@@ -248,7 +248,9 @@ describe "the class-level .config_loader attribute" do
     subject { Resque::Pool.create_configured }
 
     it "created pools use config file and hash loading logic" do
-      subject.config_loader.should be_instance_of Resque::Pool::FileOrHashLoader
+      subject.config_loader.should be_instance_of(
+        Resque::Pool::ConfigLoaders::FileOrHashLoader
+      )
     end
   end
